Extract common methods to a helpers module
